requirements.txt
Flask~=1.1.2
python-dotenv~=0.15.0
SQLAlchemy~=1.3.23
Werkzeug~=1.0.1